Fix Log on Some Platforms
On some platforms (e.g. ASR) uint32_t seems to be a long and will produce a error if %u is used. So cast it to unsinged long on all platforms and use %lu.

+++ src/app/clusters/resource-monitoring-server/resource-monitoring-server.cpp
@@ -51,8 +51,7 @@ void Instance::LoadPersistentAttributes()
         }
         else
         {
-
-            ChipLogDetail(Zcl, "ResourceMonitoring: Loaded LastChangedTime as %u", mLastChangedTime.Value());
+            ChipLogDetail(Zcl, "ResourceMonitoring: Loaded LastChangedTime as %lu", (long unsigned int) mLastChangedTime.Value()); // on some platforms uint32_t is a long, cast it to unsigned long on all platforms to prevent CI errors
         }
     }
     else
